rename and move safe_transform func, and only create it in schema init
