+++ cubedash/index/postgis/_api.py
@@ -51,7 +51,6 @@ from sqlalchemy.types import TIMESTAMP
 import cubedash.summary._schema as _schema
 from cubedash._utils import datetime_expression, default_utc
 from cubedash.index.api import EmptyDbError, ExplorerAbstractIndex
-from cubedash.index.sql import TRANSFORM_SAFE_SQL
 
 from ._schema import (  # isort: skip
     CUBEDASH_SCHEMA,
@@ -368,7 +367,7 @@ class ExplorerIndex(ExplorerAbstractIndex):
         # we use the `period_type=="all"` to get the one that covers all time.
         collection_bbox = func.Box2D(
             # TODO: reinstate ST_Transform once possible
-            func.cubedash.transform_safe(TimeOverview.footprint_geometry, 4326)
+            func.cubedash.safe_transform(TimeOverview.footprint_geometry, 4326)
         )
         bbox_array = array(
             [
@@ -438,7 +437,6 @@ class ExplorerIndex(ExplorerAbstractIndex):
         query = query.limit(limit).offset(offset)
 
         with self.index._active_connection() as conn:
-            conn.execute(text(TRANSFORM_SAFE_SQL))
             return conn.execute(query)
 
     @override
