+++ cubedash/summary/_schema.py
@@ -1,8 +1,10 @@
 from enum import Enum
 
 import structlog
+from geoalchemy2 import Geometry
 from sqlalchemy import MetaData, func, select, text
 from sqlalchemy.engine import Connection
+from sqlalchemy.sql.functions import GenericFunction
 
 _LOG = structlog.stdlib.get_logger()
 
@@ -120,5 +122,33 @@ def refresh_supporting_views(conn, concurrently: bool) -> None:
     conn.execute(
         text(f"""
     refresh materialized view {args} {CUBEDASH_SCHEMA}.mv_dataset_spatial_quality;
+    """)
+    )
+
+
+class SafeTransform(GenericFunction):
+    type = Geometry()
+    package = CUBEDASH_SCHEMA
+    identifier = "safe_transform"
+    inherit_cache = False
+    name = "safe_transform"
+
+    def __init__(self, *args, **kwargs) -> None:
+        super().__init__(*args, **kwargs)
+        self.packagenames = (CUBEDASH_SCHEMA,)
+
+
+def create_safe_transform_func(conn: Connection) -> None:
+    conn.execute(
+        text(f"""
+    create or replace function {CUBEDASH_SCHEMA}.safe_transform(geom geometry, srid int)
+    returns geometry AS $$
+    begin
+        return ST_Transform(geom, srid);
+    exception when internal_error then
+        return null;
+    end;
+    $$
+    language plpgsql;
     """)
     )
